stm32: add kernel patch to use dwc3 in dual-role mode if ALL_KMODS is set
In OpenWrt core, the package kmod-usb-dwc3 forces host mode. The dwc3 driver
with dual-role mode can be selected builtin in the target/subtarget kernel
configuration. But in case of a build with ALL_KMODS option, the kmod-dwc3
package is built, so the kernel configuration is overrided and host mode is
selected.

Add a kernel patch to disable host and gadget mode for stm32, so dual-role
is the default mode. It is a ugly hack awaiting a fix upstream in the
kmod-usb-dwc3 package.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>